Create BreakBell.bat
This simple batch lets you program a set of hours in order to play a sound, which can be whatever you want, but as it is intended for schools, I used a bell, located at C:\Bell.wav (it needs to be a .wav file). You can change the folder for whatever you like in the :bell section.

There are instructions inside the code for you to add, delete or change the set time.

The sound comes from the main output of the pc, so you will need to attach an amplifier for it to get loud enough for the whole school.